Refactor functions infrastructure
